fakefact backend/requirements.txt
Flask
flask-cors
google-generativeai
language-tool-python
Pillow
opencv-python
numpy
